+++ modules/apps/commerce/commerce-catalog-web/src/main/resources/META-INF/resources/init.jsp
@@ -28,8 +28,8 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
 <%@ page import="com.liferay.commerce.catalog.web.internal.constants.CommerceCatalogFDSNames" %><%@
+page import="com.liferay.commerce.catalog.web.internal.constants.CommerceCatalogScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.catalog.web.internal.display.context.CommerceCatalogDisplayContext" %><%@
-page import="com.liferay.commerce.catalog.web.internal.servlet.taglib.ui.constants.CommerceCatalogScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.currency.model.CommerceCurrency" %><%@
 page import="com.liferay.commerce.price.list.constants.CommercePriceListConstants" %><%@
 page import="com.liferay.commerce.price.list.exception.NoSuchPriceListException" %><%@
